Split extract_landmarks into non-ROS (base) and ROS part
